Add CheckStyle
