+++ src/main/java/alex/Alex.java
@@ -2,20 +2,29 @@ package alex;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
-
 import java.time.format.DateTimeParseException;
 
 import alex.command.Command;
-import alex.task.Pair;
+import alex.task.TaskList;
 
 /**
- * Represents the chatbot Alex
+ * Represents the chatbot Alex which handles user interactions, processes commands, and manages tasks.
  */
 public class Alex {
     private Storage storage;
     private TaskList tasks;
     private Ui ui;
 
+    /**
+     * Constructs an instance of the Alex chatbot.
+     * Initializes the user interface, storage system, and task list.
+     * Loads existing tasks from the specified file path.
+     *
+     * @param filePath The path to the file where tasks are stored.
+     * @throws AlexException If there is an error specific to the Alex application.
+     * @throws FileNotFoundException If the file specified by filePath is not found.
+     * @throws DateTimeParseException If there is an error parsing the date in the task list.
+     */
     public Alex(String filePath) {
         this.ui = new Ui();
         this.storage = new Storage(filePath);
@@ -27,7 +36,11 @@ public class Alex {
     }
 
     /**
-     * Generates a response for the user's chat message.
+     * Processes the user's input and generates a response based on the command.
+     * Executes the command and returns a pair containing the response and the command type.
+     *
+     * @param input The user's chat message.
+     * @return A pair where the first element is the response to the user's input and the second element is the command type.
      */
     public Pair<String, String> getResponse(String input) {
         try {
