+++ src/main/java/alex/command/DeleteCommand.java
@@ -1,20 +1,24 @@
 package alex.command;
 
+import java.io.IOException;
 import java.util.Scanner;
 
-import java.io.IOException;
-
-import alex.TaskList;
-import alex.Ui;
-import alex.Storage;
 import alex.AlexException;
+import alex.Storage;
+import alex.Ui;
+import alex.task.TaskList;
 
 /**
- * Represents the command by user to delete a Task from Tasklist.
+ * Represents the command by user to delete a Task from TaskList.
  */
 public class DeleteCommand extends Command {
     private Scanner lineScanner;
 
+    /**
+     * Constructs a DeleteCommand instance.
+     *
+     * @param lineScanner Scanner object used to read user input.
+     */
     public DeleteCommand(Scanner lineScanner) {
         this.lineScanner = lineScanner;
     }
@@ -22,35 +26,36 @@ public class DeleteCommand extends Command {
     /**
      * {@inheritDoc}
      *
-     * Deletes the Task from TaskList, save the changes to file and informs user.
+     * Deletes the Task from TaskList, saves the changes to file, and informs the user.
      *
-     * @param tasks Tasklist that holds the list of Tasks.
-     * @param ui Ui object that displays messages to user based on action taken by chatbot.
-     * @param storage Storage object that saves changes to file.
-     * @throws AlexException If user provides wrong input such as not providing an integer number or
+     * @param tasks TaskList that holds the list of Tasks.
+     * @param ui Ui object that displays messages to the user based on the action taken by the chatbot.
+     * @param storage Storage object that saves changes to the file.
+     * @return A message indicating the result of the delete operation.
+     * @throws AlexException If the user provides incorrect input, such as not providing an integer number or
      * providing a number that is too small or too big.
-     * @throws IOException If there are issues saving the changes to file.
+     * @throws IOException If there are issues saving the changes to the file.
      */
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) throws AlexException, IOException {
         if (!lineScanner.hasNext()) {
-            throw new AlexException("Oh no! Please provide an integer number after 'delete' indicating the " +
-                    "task number to delete!");
+            throw new AlexException("Oh no! Please provide an integer number after 'delete' indicating the "
+                    + "task number to delete!");
         }
         String taskNumberStr = lineScanner.next();
-        int taskNumber = 0;
+        int taskNumber;
 
-        //handles exception where user write too much
+        // Handles case where user writes too much
         if (lineScanner.hasNext()) {
             throw new AlexException("Wait! Please only provide a number after 'delete'!");
         }
 
-        //handles case where user doesn't provide a number or not an integer
+        // Handles case where user doesn't provide a number or provides a non-integer
         try {
             taskNumber = Integer.valueOf(taskNumberStr);
         } catch (NumberFormatException e) {
-            throw new AlexException("Oh no! Please only provide an integer number after 'delete' indicating\n" +
-                    "the task number to delete!");
+            throw new AlexException("Oh no! Please only provide an integer number after 'delete' indicating\n"
+                    + "the task number to delete!");
         }
 
         if (taskNumber < 1 || taskNumber > tasks.getSize()) {
@@ -63,7 +68,7 @@ public class DeleteCommand extends Command {
     /**
      * {@inheritDoc}
      *
-     * @return false as user is not done yet.
+     * @return false as the user is not done yet.
      */
     @Override
     public boolean isExit() {
@@ -75,3 +80,4 @@ public class DeleteCommand extends Command {
         return "DeleteCommand";
     }
 }
+
