+++ src/main/java/alex/javafx/Main.java
@@ -3,14 +3,10 @@ package alex.javafx;
 import java.io.IOException;
 
 import alex.Alex;
-import alex.Ui;
 import javafx.application.Application;
-import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
-import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 
 /**
