+++ src/main/java/alex/task/Deadline.java
@@ -8,7 +8,15 @@ import java.time.format.DateTimeFormatter;
  */
 public class Deadline extends Task {
     private LocalDateTime deadline;
-    public Deadline (String taskName, boolean isCompleted, LocalDateTime deadline) {
+
+    /**
+     * Constructs a Deadline Task with the specified details.
+     *
+     * @param taskName The name or description of the task.
+     * @param isCompleted A boolean indicating whether the task is completed.
+     * @param deadline The deadline for the task as a LocalDateTime object.
+     */
+    public Deadline(String taskName, boolean isCompleted, LocalDateTime deadline) {
         super(taskName, isCompleted);
         this.deadline = deadline;
     }
@@ -16,8 +24,10 @@ public class Deadline extends Task {
     /**
      * {@inheritDoc}
      *
-     * Returns the String representation of the Task to be displayed to user
-     * that includes its completion status, description and deadline.
+     * Returns the String representation of the Task to be displayed to the user
+     * that includes its completion status, description, and deadline.
+     *
+     * @return A String representation of the Deadline Task.
      */
     @Override
     public String toString() {
@@ -25,9 +35,16 @@ public class Deadline extends Task {
                 + this.deadline.format(DateTimeFormatter.ofPattern("MMM d yyyy h.mma")) + ")";
     }
 
+    /**
+     * Returns a String representation of the Deadline Task suitable for storage.
+     * The format includes the task details and the deadline in "yyyy-MM-dd HHmm" format.
+     *
+     * @return A String representation of the Deadline Task for storage.
+     */
     @Override
     public String toStorageString() {
         return "[D]" + super.toString() + " /by "
                 + this.deadline.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"));
     }
 }
+
