+++ src/main/java/alex/task/Task.java
@@ -1,13 +1,18 @@
 package alex.task;
 
 /**
- * Represents the task that the user wrote.
- * Contains a description and a completion status.
+ * Represents a task with a description and a completion status.
  */
 public class Task {
     private String taskName;
-
     private boolean isCompleted;
+
+    /**
+     * Constructs a Task with the specified description and completion status.
+     *
+     * @param taskName The name or description of the task.
+     * @param isCompleted A boolean indicating whether the task is completed.
+     */
     public Task(String taskName, boolean isCompleted) {
         this.taskName = taskName;
         this.isCompleted = isCompleted;
@@ -28,32 +33,37 @@ public class Task {
     }
 
     /**
-     * Returns the String representation of the Task to be displayed to user
-     * that includes its completion status and description.
+     * Returns the String representation of the Task to be displayed to the user,
+     * including its completion status and description.
+     *
+     * @return A String representation of the Task.
      */
     @Override
     public String toString() {
-        String box = "";
-        if (this.isCompleted) {
-            box = "[X]";
-        } else {
-            box = "[ ]";
-        }
-
+        String box = this.isCompleted ? "[X]" : "[ ]";
         return String.format("%s %s", box, this.taskName);
     }
 
     /**
-     * Returns the String representation of the Task to be saved in the file
-     * and retrieved when the chatbot is run again, which is similar to format of user input. The 2
-     * different String representation formats allow reuse of Parser functions to convert the text in
-     * the file back to Task objects when the file is loaded at the start of every run.
+     * Returns the String representation of the Task to be saved in a file.
+     * This format is similar to the format of user input to facilitate
+     * conversion of text from the file back to Task objects when the file
+     * is loaded at the start of each run.
+     *
+     * @return A String representation of the Task for storage.
      */
     public String toStorageString() {
         return "";
     }
 
-    public boolean hasSearchString(String searchString) {
+    /**
+     * Checks if the task description contains the specified search string.
+     *
+     * @param searchString The string to search for within the task description.
+     * @return True if the task description contains the search string, false otherwise.
+     */
+    public boolean checkHaveSearchString(String searchString) {
         return this.taskName.contains(searchString);
     }
 }
+
