+++ src/test/java/alex/task/DeadlineTest.java
@@ -1,10 +1,11 @@
 package alex.task;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import java.time.LocalDateTime;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 public class DeadlineTest {
     @Test
     public void testToStringMethod() {
