+++ src/test/java/alex/task/TodoTest.java
@@ -1,8 +1,8 @@
 package alex.task;
 
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 
 public class TodoTest {
     @Test
